fix: code enhancements
