+++ src/app/(app)/layout.tsx
@@ -1,16 +1,29 @@
-import { Inter } from 'next/font/google'
-import './globals.css'
+import { Inter } from "next/font/google";
 
-const inter = Inter({ subsets: ['latin'] })
+const inter = Inter({ subsets: ["latin"] });
 
-export default function AppLayout({
-  children,
-}: {
-  children: React.ReactNode
-}) {
+import { AppSidebar } from "@/components/app-sidebar";
+
+import {
+  SidebarInset,
+  SidebarProvider,
+} from "@/components/ui/sidebar";
+
+export default function AppLayout({ children }: { children: React.ReactNode }) {
   return (
     <html lang="pt-br">
-      <body className={inter.className}>{children}</body>
+      <body className={inter.className}>
+        <SidebarProvider
+          style={
+            {
+              "--sidebar-width": "350px",
+            } as React.CSSProperties
+          }
+        >
+          <AppSidebar />
+          <SidebarInset>{children}</SidebarInset>
+        </SidebarProvider>
+      </body>
     </html>
-  )
+  );
 }
