--- src/app/(auth)/login/page.tsx
@@ -1,5 +1,0 @@
-import { LoginForm } from "@/components/login-form"
-
-export default function Page() {
-  return <LoginForm />;
-}
