+++ src/app/(auth)/signin/page.tsx
@@ -1,0 +1,5 @@
+import { SigninForm } from "@/components/signin-form";
+
+export default function Page() {
+  return <SigninForm />;
+}
